Split date workflow into separate modules
This commit splits the date workflow into separate modules
that allow extensions, and also adds a `common` module
interface for them with types to reuse.
